% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmose_extras.R
\name{initialize_osmose}
\alias{initialize_osmose}
\title{Create initialization file for an OSMOSE configuration}
\usage{
initialize_osmose(input, output = NULL, ...)
}
\arguments{
\item{input}{Filename of the main configuration file}

\item{output}{Ouput file containing the initialization configuration}

\item{...}{}
}
\value{

}
\description{
Create initialization file for an OSMOSE configuration
}
